feat: Add Yarn command to run Native Android tests
